requirements.txt
aliyun-python-sdk-dysmsapi==1.0.0
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.0
django-shortuuidfield==0.1.3
idna==2.7
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
shortuuid==0.5.0
six==1.11.0
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
